add a bunch of new stuff

+++ src/components/node/NodeEditor.vue
@@ -1,39 +1,38 @@
 <template>
   <div class="node-editor editor-pane" tabindex="-1">
     <shortcut-receiver @shortcut="handleShortcut" style="height: 100%;">
-      <el-container style="height: 100% !important;">
-        <el-header style="display: flex; align-items: center; border-bottom: 1px solid var(--el-border-color)">
-          <el-breadcrumb>
-            <el-breadcrumb-item @click="ctx.activePath.value = rootPath">
-              Storyboard
-            </el-breadcrumb-item>
-            <el-breadcrumb-item v-for="(path) in ctx.activePath.value.parentPaths" :key="path.toString()"
-                                @click="ctx.activePath.value = path">
-              {{ path.current }}
-            </el-breadcrumb-item>
-          </el-breadcrumb>
-        </el-header>
-        <el-main style="padding:0">
+      <div style="height: 100% !important;">
+        <node-editor-header @createNode="createNode"/>
+        <div>
           <div ref="nodePane"
                class="node-pane"
-               @mousedown.left="handleDragSelect"
-               @mousedown.middle="handlePan"
-               @click.left="deselectAll"
+               @mousedown.left.prevent="handleDragSelect"
+               @mousedown.middle.prevent="handlePan"
                @wheel="handleScroll">
             <svg class="connections-container" v-if="initialized">
-              <g v-for="connection in nodeSystem.connectionList" :key="connection.id">
-                <node-connection-view :from="connection.from" :to="connection.to" :editor-id="id" :bounds="bounds"
-                                      :transform="stack.current"
-                                      :circular="connection.circular.value"
-                                      v-if="connection !== hiddenConnection"
-                />
-              </g>
+              <g :transform="connectionsTransform">
+                <g v-for="connection in connections" :key="connection.id">
+                  <node-connection-view :connection="connection"
+                                        :from="connection.from"
+                                        :to="connection.to"
+                                        :editor-id="id"
+                                        :bounds="bounds"
+                                        :transform="transform"
+                                        :circular="connection.circular.value"
+                                        v-if="connection !== hiddenConnection"
+                                        @update:hovering="handleNodeHoveringChanged(connection, $event)"
+                  />
+                </g>
 
 
-              <g v-if="tempConnection.active">
-                <node-connection-view :from="tempConnection.from" :to="tempConnection.to" :editor-id="id"
-                                      :bounds="bounds"
-                                      :transform="stack.current"/>
+                <g v-if="tempConnection.active">
+                  <node-connection-view :from="tempConnection.from"
+                                        :to="tempConnection.to"
+                                        :editor-id="id"
+                                        :bounds="bounds"
+                                        :transform="transform"
+                  />
+                </g>
               </g>
             </svg>
 
@@ -44,8 +43,9 @@
                          :selecting="selecting"
                          :selection-candidate="isSelectionCandidate(node)"
                          :ctx="ctx"
-                         :key="node.name"
-                         :viewport-scale="stack.current.scale"
+                         :key="node.name.value"
+                         :viewport-scale="transform.scale"
+                         :hovered-connections="hoveredConnections"
                          @connection:start="startConnection"
                          @socket:enter="hoveringSocket = $event"
                          @socket:leave="hoveringSocket = undefined"
@@ -53,41 +53,72 @@
               <select-box v-if="selecting" :from="selectFrom" :to="selectTo"/>
             </div>
           </div>
-        </el-main>
-      </el-container>
+        </div>
+        <add-node-view/>
+      </div>
     </shortcut-receiver>
   </div>
 </template>
 
 <script setup lang="ts">
 import NodeView from './NodeView.vue';
-import {computed, onBeforeUnmount, onMounted, reactive, ref, shallowReactive, shallowRef} from "vue";
+import {
+  computed,
+  onBeforeUnmount,
+  onMounted,
+  reactive,
+  ref,
+  shallowReactive,
+  shallowRef,
+  watch,
+  watchPostEffect
+} from "vue";
 import {drag} from "@/util/event";
-import {EditorStack} from "@/components/node/util";
+import {EditorLocation} from "@/components/node/util";
 import {useContext} from "@/editor/ctx/use";
 import {Vec2} from "@/util/math";
 import SelectBox from "@/components/SelectBox.vue";
 import NodeConnectionView from "@/components/node/NodeConnectionView.vue";
-import {NodeSystem} from "@/editor/node/system";
 import {v4 as uuid} from 'uuid'
-import {Node, NodeInput, NodeOutput} from "@/editor/node";
+import {Node} from "@/editor/node";
 import ShortcutReceiver from "@/components/ShortcutReceiver.vue";
-import {AddConnectionCommand, DeleteNodesCommand, RemoveConnectionCommand} from "@/editor/ctx/command/node";
+import {
+  AddConnectionCommand,
+  CreateNodeCommand,
+  DeleteNodesCommand,
+  LayoutNodesCommand,
+  RemoveConnectionCommand
+} from "@/editor/ctx/command/node";
 import {NodeConnection} from "@/editor/node/connection";
-import {NodePath} from "@/editor/node/path";
+import {NodeInput, NodeOutput} from "@/editor/node/input";
+import {EditorCommandCollection} from "@/editor/ctx/editorCommand";
+import gsap from "gsap";
+import AddNodeView from "@/components/node/AddNodeView.vue";
+import NodeEditorHeader from "@/components/node/NodeEditorHeader.vue";
+import Tween = gsap.core.Tween;
 
-const stack = reactive(new EditorStack())
+//const stack = reactive(new EditorStack())
 
 const ctx = useContext()
 
-const nodeSystem = computed(() =>
-    ctx.getObject(ctx.activePath.value) as NodeSystem<any>
-  )
+const locations = ctx.locations
 
-const rootPath = NodePath.root()
+const currentLocation = computed(() => {
+  const path = ctx.activeNodeSystem.value.path.toString()
+  if (!locations.has(path)) {
+    locations.set(path, new EditorLocation(path))
+  }
+  return locations.get(path)!
+})
+
+
+const nodeSystem = ctx.activeNodeSystem
+
+const connections = shallowRef<ReadonlyArray<NodeConnection<any>>>([])
+
+watchPostEffect(() => connections.value = nodeSystem.value.connectionList)
 
 const nodePane = ref<HTMLDivElement>()
-
 const id = uuid().substring(0, 4)
 const initialized = ref(false)
 const bounds = shallowRef<DOMRect>()
@@ -95,6 +126,8 @@ const bounds = shallowRef<DOMRect>()
 function onResize() {
   bounds.value = nodePane.value?.getBoundingClientRect()
 }
+
+const hoveredConnections = shallowReactive(new Set<NodeConnection>())
 
 const observer = new ResizeObserver(onResize)
 
@@ -105,32 +138,59 @@ onMounted(() => {
   observer.observe(nodePane.value!)
 })
 
+const transition = reactive({
+  scale: 1
+})
+
+const transform = computed(() => {
+  const location = currentLocation.value.clone()
+  location.scale *= transition.scale
+  location.position.move(new Vec2(
+      -(1 - 1 / transition.scale) * (bounds.value?.width ?? 0) / 2,
+      -(1 - 1 / transition.scale) * (bounds.value?.height ?? 0) / 2,
+  ))
+  return location
+})
+
+let tween: Tween | undefined;
+
+watch(nodeSystem, (system, old) => {
+  tween?.kill()
+
+  if (system.isChildOf(old)) {
+    transition.scale = 0.75
+    tween = gsap.to(transition, {duration: 0.3, scale: 1})
+  } else if (old.isChildOf(system)) {
+    transition.scale = 1.5
+    tween = gsap.to(transition, {duration: 0.3, scale: 1})
+  }
+})
+
 onBeforeUnmount(() => {
   observer.disconnect()
 })
 
-
 function handleScroll(evt: WheelEvent) {
   let scrollAmount = evt.deltaY;
 
-  const newScale = stack.current.scale * (1 - scrollAmount / 3000);
+  const newScale = currentLocation.value.scale * (1 - scrollAmount / 3000);
 
   const currentPoint = new Vec2(
-      evt.offsetX / stack.current.scale - stack.current.position.x,
-      evt.offsetY / stack.current.scale - stack.current.position.y,
+      evt.offsetX / currentLocation.value.scale - currentLocation.value.position.x,
+      evt.offsetY / currentLocation.value.scale - currentLocation.value.position.y,
   );
 
-  const newPoint = new Vec2(evt.offsetX / newScale - stack.current.position.x, evt.offsetY / newScale - stack.current.position.y)
+  const newPoint = new Vec2(evt.offsetX / newScale - currentLocation.value.position.x, evt.offsetY / newScale - currentLocation.value.position.y)
   const diff = newPoint.sub(currentPoint)
-  stack.current.position.move(diff)
-  stack.current.scale = newScale;
+  currentLocation.value.position.move(diff)
+  currentLocation.value.scale = newScale;
 }
 
 function handlePan(evt: MouseEvent) {
   drag(evt, {
     onDrag({delta}) {
-      stack.current.position =
-          stack.current.position.add(delta.divF(stack.current.scale))
+      currentLocation.value.position =
+          currentLocation.value.position.add(delta.divF(currentLocation.value.scale))
     }
   })
 }
@@ -153,8 +213,8 @@ function handleDragSelect(evt: MouseEvent) {
     el: nodePane.value,
     onDragStart({start}) {
       const pos = new Vec2(
-          start.x / stack.current.scale - stack.current.position.x,
-          start.y / stack.current.scale - stack.current.position.y,
+          start.x / currentLocation.value.scale - currentLocation.value.position.x,
+          start.y / currentLocation.value.scale - currentLocation.value.position.y,
       );
       selectFrom.value = pos
       selectTo.value = pos
@@ -162,8 +222,8 @@ function handleDragSelect(evt: MouseEvent) {
     },
     onDrag({current}) {
       selectTo.value = new Vec2(
-          current.x / stack.current.scale - stack.current.position.x,
-          current.y / stack.current.scale - stack.current.position.y,
+          current.x / currentLocation.value.scale - currentLocation.value.position.x,
+          current.y / currentLocation.value.scale - currentLocation.value.position.y,
       )
     },
     onDragEnd() {
@@ -172,32 +232,55 @@ function handleDragSelect(evt: MouseEvent) {
           additive
       )
       selecting.value = false
+    },
+    onClick: () => {
+      nodeSystem.value.select([])
     }
   })
-}
-
-
-function deselectAll() {
-  nodeSystem.value.select([])
 }
 
 const styles = computed(() =>
     ({
       "transform-origin": "0 0",
-      "transform": `scale(${stack.current.scale})`,
-      "top": stack.current.scale * stack.current.position.y + "px",
-      "left": stack.current.scale * stack.current.position.x + "px"
+      "transform": `scale(${transform.value.scale})`,
+      "top": transform.value.scale * transform.value.position.y + "px",
+      "left": transform.value.scale * transform.value.position.x + "px"
     })
 );
 
-function handleShortcut(shortcut: string) {
-  switch (shortcut) {
-    case 'delete': {
+const connectionsTransform = computed(() => `translate(${transform.value.scale * transform.value.position.x}, ${transform.value.scale * transform.value.position.y}) scale(${transform.value.scale})`)
+
+function handleShortcut(evt: CustomEvent) {
+  switch (evt.detail as string) {
+    case 'delete':
       if (ctx.activeNode.value)
+        if (nodeSystem.value.selectedNodes.length > 0)
+          ctx.executeCommand(
+              new DeleteNodesCommand(ctx, nodeSystem.value.selectedNodes.map(it => it.path))
+          )
+      if (nodeSystem.value.selectedConnections.length > 0)
         ctx.executeCommand(
-            new DeleteNodesCommand(ctx, nodeSystem.value.selectedNodes.map(it => it.path))
+            new EditorCommandCollection(ctx, 'Delete Connections', nodeSystem.value.selectedConnections.map(connection =>
+                new RemoveConnectionCommand(ctx, connection.from.node.path, connection.from.index, connection.to.node.path, connection.to.index)
+            ))
         )
-    }
+      break;
+    case 'l':
+      console.log('layout nodes')
+      if (nodeSystem.value.selectedNodes.length) {
+        ctx.executeCommand(new LayoutNodesCommand(ctx, nodeSystem.value.path, nodeSystem.value.selectedNodes.map(it => it.name.value)))
+      } else {
+        ctx.executeCommand(new LayoutNodesCommand(ctx, nodeSystem.value.path, [...nodeSystem.value.nodeNames]))
+      }
+      break;
+    case 'ctrl+a':
+      evt.preventDefault()
+      nodeSystem.value.select([...nodeSystem.value.nodeList])
+      break;
+    case 'ctrl+d':
+      evt.preventDefault()
+      nodeSystem.value.select([])
+      break;
   }
 }
 
@@ -233,6 +316,11 @@ function startConnection({evt, socket}: { socket: NodeInput | NodeOutput, evt: M
       }
     },
     onDrag({current}) {
+      current = current.clone()
+          .divF(transform.value.scale)
+          .sub(transform.value.position)
+
+
       if (socket instanceof NodeInput) {
         if (hoveringSocket.value instanceof NodeOutput)
           tempConnection.from = hoveringSocket.value
@@ -277,6 +365,32 @@ function startConnection({evt, socket}: { socket: NodeInput | NodeOutput, evt: M
   })
 }
 
+function createNode(type: string) {
+  const rect = nodePane.value!.getBoundingClientRect()
+  ctx.executeCommand(new CreateNodeCommand(ctx, nodeSystem.value.path, type, {
+    position: new Vec2(
+        -currentLocation.value.position.x + rect.width / 2 / currentLocation.value.scale,
+        -currentLocation.value.position.y + rect.height / 2 / currentLocation.value.scale
+    )
+  }))
+}
+
+function handleNodeHoveringChanged(connection: NodeConnection, hovering: boolean) {
+  if (hovering)
+    hoveredConnections.add(connection)
+  else
+    hoveredConnections.delete(connection)
+}
+
+watch(connections, () => {
+  const previous = new Set(hoveredConnections)
+  hoveredConnections.clear()
+  connections.value.forEach(it => {
+    if (previous.has(it))
+      hoveredConnections.add(it)
+  })
+})
+
 </script>
 
 
@@ -294,6 +408,7 @@ function startConnection({evt, socket}: { socket: NodeInput | NodeOutput, evt: M
   position: absolute;
   width: 100%;
   height: 100%;
+  pointer-events: none;
 }
 
 .connections-container {
